fix import loop
