+++ src/neuroglancer_annotation_ui/statebuilder/utils.py
@@ -1,6 +1,4 @@
 from collections.abc import Iterable
-from annotationframeworkclient import FrameworkClient
-from .statebuilder import ImageLayerConfig, SegmentationLayerConfig
 import numpy as np
 import pandas as pd
 
@@ -36,47 +34,3 @@ def bucket_of_values(col, data, item_is_array=False, array_width=3):
             return np.concatenate(dataseries.values)
         else:
             return dataseries.values
-
-
-def sources_from_infoclient(dataset_name, segmentation_type='graphene', image_kws={}, segmentation_kws={}, client=None):
-    """Generate an Image and Segmentation source from a dataset name.
-
-    Parameters
-    ----------
-    dataset_name : str
-        Dataset name for client
-    segmentation_type : 'graphene' or 'flat', optional
-        Which segmentation type for try first, graphene or flat. It will fall back to the other if not found. By default 'graphene'
-    image_kws : dict, optional
-        Keyword arguments for an ImageLayerConfig (other than source), by default {}
-    segmentation_kws : dict, optional
-        Keyword arguments for a SegmentationLayerConfig (other than source), by default {}
-    client : InfoClient or None, optional
-        Predefined info client for lookup
-
-    Returns
-    -------
-    ImageLayerConfig
-        Config for an image layer in the statebuilder 
-    SegmentationLayerConfig
-        Config for a segmentation layer in the statebuilder
-    """
-
-    if client is None:
-        client = FrameworkClient(dataset_name=dataset_name)
-
-    image_source = client.info.image_source(format_for='neuroglancer')
-    if segmentation_type == "graphene":
-        seg_source = client.info.graphene_source(format_for='neuroglancer')
-        if seg_source is None:
-            seg_source = client.info.flat_segmentation_source(
-                format_for='neuroglancer')
-    else:
-        seg_source = client.info.flat_segmentation_source(
-            format_for='neuroglancer')
-        if seg_source is None:
-            seg_source = client.info.graphene_source(format_for='neuroglancer')
-
-    image_layer = ImageLayerConfig(image_source, *image_kws)
-    seg_layer = SegmentationLayerConfig(seg_source, *segmentation_kws)
-    return image_layer, seg_layer
